Test jtimer package
Moved all packages to a new pattern of starting with `j`. This is because the intended use of the packages will be to rename the package to an alias of "j-something" in the first place, to avoid collisions with other more well known packages. So why not just start it out that way in the first place.

Add jerrors package.

Generate fakes for jtimer package.

Add tools.go for tracking counterfeiter version